+++ example/lib/searching_example.dart
@@ -2,23 +2,26 @@ import 'package:algorithmic/algorithmic.dart';
 
 void main() {
   // Linear searching
-  final arr = [1, -1, 2, 2, -2];
+  final arr = [-2, -1, 1, 2, 2, 3];
 
-  print("find 2 in $arr [expects 2]: "
+  print("linear search 2 in $arr [expects 3]: "
       "${linearSearch(arr, 2)}");
+      
+  print("binary search 2 in $arr [expects 3]: "
+      "${binarySearch(arr, 2)}");
 
-  print("find 2 in $arr with offset = 3 [expects 3]: "
-      "${linearSearch(arr, 2, offset: 3)}");
+  print("binary search 2 in $arr with offset = 3 [expects 3]: "
+      "${binarySearch(arr, 2, start: 3)}");
 
-  print("find 2 in $arr with count = 2 [expects -1]: "
-      "${linearSearch(arr, 2, count: 2)}");
+  print("binary search 2 in $arr with count = 2 [expects -1]: "
+      "${binarySearch(arr, 2, count: 2)}");
 
-  print("find 2 in $arr with offset = 2 and count = 1 [expects 2]: "
-      "${linearSearch(arr, 2, offset: 2, count: 1)}");
+  print("binary search 2 in $arr with offset = 2 and count = 1 [expects 4]: "
+      "${binarySearch(arr, 2, start: 4, count: 1)}");
 
-  print("find 2 in $arr with custom comparator [expects 2]: "
-      "${linearSearch(arr, 2, compare: (int a, int b) => (a - b) * (a - b))}");
+  print("linear search 2 in $arr with custom comparator [expects 0]: "
+      "${linearSearchBy(arr, (int e) => e * e == 4)}");
 
-  print("find 2 in $arr reversed [expects 3]: "
-      "${linearSearchReversed(arr, 2)}");
+  print("binary search maximum index of 2 in $arr reversed [expects 4]: "
+      "${binarySearchMax(arr, 2)}");
 }