+++ lib/src/searching/binary.dart
@@ -18,7 +18,9 @@ import 'package:algorithmic/src/utils/comparators.dart';
 /// * If [start] is not below the length of the [list], the length is returned.
 /// * If the [count] parameter is given, it will check up to [count] numbers of items.
 /// * The [count] must not be negative. Otherwise, [RangeError] is thrown.
-/// * To perform a binary search using this method, pass [exactMatch] to true.
+/// * [compare] is a custom comparator function between a list element and the value.
+///   If it is null, `compareTo` method of [list] item is used. On [TypeError], -1 is returned.
+/// * You can [exactMatch] to true if you want to perform a binary search.
 ///
 /// ## Details
 ///
@@ -27,7 +29,9 @@ import 'package:algorithmic/src/utils/comparators.dart';
 /// If the middle item of the range is less than the [value], the right half of
 /// the range will be selected, otherwise the right half. After this process is
 /// done, we are left with a singular range containing only one item.
-/// The index of this item will be returned at the end.
+///
+/// if [exactMatch] is false, the index of this item will be returned at the end.
+/// Otherwise, the if the item is equal to [value], the index will be returned or -1.
 ///
 /// -------------------------------------------------------------------------
 /// Complexity: Time `O(log n)` | Space `O(1)`
@@ -36,6 +40,7 @@ int lowerBound<E, V>(
   final V value, {
   final int? start,
   final int? count,
+  final EntryComparator<E, V>? compare,
   final bool exactMatch = false,
 }) {
   int i, j, l, r, m, c;
@@ -55,19 +60,37 @@ int lowerBound<E, V>(
 
   l = i;
   r = j;
-  while (l < r) {
-    // the middle of the range
-    m = l + ((r - l) >> 1);
+  if (compare == null) {
+    while (l < r) {
+      // the middle of the range
+      m = l + ((r - l) >> 1);
 
-    // compare middle item with value
-    c = (list[m] as Comparable).compareTo(value);
+      // compare middle item with value
+      c = (list[m] as Comparable).compareTo(value);
 
-    if (c < 0) {
-      // middle item is lesser, select right range
-      l = m + 1;
-    } else {
-      // middle item is either equal or greater, select left range
-      r = m;
+      if (c < 0) {
+        // middle item is lesser, select right range
+        l = m + 1;
+      } else {
+        // middle item is either equal or greater, select left range
+        r = m;
+      }
+    }
+  } else {
+    while (l < r) {
+      // the middle of the range
+      m = l + ((r - l) >> 1);
+
+      // compare middle item with value
+      c = compare(list[m], value);
+
+      if (c < 0) {
+        // middle item is lesser, select right range
+        l = m + 1;
+      } else {
+        // middle item is either equal or greater, select left range
+        r = m;
+      }
     }
   }
 
@@ -79,88 +102,12 @@ int lowerBound<E, V>(
   }
 
   // prepare binary search result
-  if (i <= l && l < j && list[l] == value) {
-    return l;
-  }
-  return -1;
-}
-
-/// Returns the index of the _first_ item from a sorted [list], for which the
-/// [test] method return `false`, otherwise if it returns `true` for all items,
-/// the length of the [list] is returned.
-///
-/// ## Parameters
-///
-/// * The [list] must be a sorted list of items, otherwise the behavior of this
-///   method is not defined.
-/// * If [test] method should be consitent for every items on the list. For exmaple,
-///   if `list[i] == list[j]` is true, then `test(list[i]) == test(list[j])` must
-///   also be true.
-/// * If [start] is given, search start there and go towards the end of the [list].
-/// * If [start] is not below the length of the [list], the length is returned.
-/// * If [start] is negative, search starts at [start] + [count] or 0, whichever is greater.
-/// * If the [count] parameter is given, it will check up to [count] numbers of items.
-/// * The [count] must not be negative, otherwise [RangeError] is thrown.
-/// * To perform a binary search using this method, pass the [testEqual] function.
-///
-/// ## Details
-///
-/// The search will begin with a range from [start] and consider at most [count]
-/// number of items. In each iteration, the range will be narrowed down by half.
-/// If [test] returns `true` for the item at the middle of the range, right half
-/// of the range will be selected, otherwise the left half. After this process is
-/// done, we are left with a singular range containing only one item.
-/// The index of this item will be returned at the end.
-///
-/// -------------------------------------------------------------------------
-/// Complexity: Time `O(log n)` | Space `O(1)`
-int lowerBoundBy<E>(
-  final List<E> list,
-  final LessThanTest<E> test, {
-  final int? start,
-  final int? count,
-  final EqualityTest<E>? testEqual,
-}) {
-  int i, j, l, r, m;
-  final int n = list.length;
-
-  // determine range [i, j)
-  i = start ?? 0;
-  j = n;
-  if (count != null) {
-    if (count < 0) {
-      throw RangeError("count can not be negative");
-    }
-    j = i + count;
-    if (j > n) j = n;
-  }
-  if (i < 0) i = 0;
-
-  l = i;
-  r = j;
-  while (l < r) {
-    // the middle of the range
-    m = l + ((r - l) >> 1);
-
-    if (test(list[m])) {
-      // middle item is lesser, select right range
-      l = m + 1;
+  if (i <= l && l < j) {
+    if (compare == null) {
+      if (list[l] == value) return l;
     } else {
-      // middle item is either equal or greater, select left range
-      r = m;
+      if (compare(list[l], value) == 0) return l;
     }
-  }
-
-  if (l > j) l = j;
-
-  if (testEqual == null) {
-    // lower bound index
-    return l;
-  }
-
-  // prepare binary search result
-  if (i <= l && l < j && testEqual(list[l])) {
-    return l;
   }
   return -1;
 }
@@ -183,9 +130,7 @@ int lowerBoundBy<E>(
 ///
 /// ## Details
 ///
-/// If [compare] is null, the [lowerBound] is called with `exactMatch` as true, otherwise the
-/// [lowerBoundBy] is called with a `testEqual` function to perform binary search.
-/// On [TypeError] or [RangeError], -1 is returned.
+/// Internally it calls the [lowerBound] method, passing true to the `exactMatch` parameter.
 ///
 /// -------------------------------------------------------------------------
 /// Complexity: Time `O(log n)` | Space `O(1)`
@@ -196,30 +141,14 @@ int binarySearch<E, V>(
   final int? count,
   final EntryComparator<E, V>? compare,
 }) {
-  try {
-    if (compare == null) {
-      return lowerBound<E, V>(
-        list,
-        value,
-        start: start,
-        count: count,
-        exactMatch: true,
-      );
-    } else {
-      return lowerBoundBy<E>(
-        list,
-        (e) => compare(e, value) < 0,
-        start: start,
-        count: count,
-        testEqual: (e) => compare(e, value) == 0,
-      );
-    }
-  } on RangeError {
-    // ignore: empty_catches
-  } on TypeError {
-    // ignore: empty_catches
-  }
-  return -1;
+  return lowerBound<E, V>(
+    list,
+    value,
+    start: start,
+    count: count,
+    compare: compare,
+    exactMatch: true,
+  );
 }
 
 /// Returns the index of the _first_ item from a sorted [list] that is strictly
@@ -243,8 +172,14 @@ int binarySearch<E, V>(
 ///
 /// ## Details
 ///
-/// Internally the [lowerBoundBy] is called comparing whether the list item is less
-/// than or equal to the value as [test] function.
+/// The search will begin with a range from [start] and consider at most [count]
+/// number of items. In each iteration, the range will be narrowed down by half.
+/// If the middle item of the range is less or equal to the [value], right half
+/// of the range will be selected, otherwise the right half. After this process
+/// is done, we are left with a singular range containing only one item.
+///
+/// if [exactMatch] is false, the index of this item will be returned at the end.
+/// Otherwise, the if the item is equal to [value], the index will be returned or -1.
 ///
 /// -------------------------------------------------------------------------
 /// Complexity: Time `O(log n)` | Space `O(1)`
@@ -254,22 +189,76 @@ int upperBound<E, V>(
   final int? start,
   final int? count,
   final EntryComparator<E, V>? compare,
+  final bool exactMatch = false,
 }) {
-  if (compare == null) {
-    return lowerBoundBy<E>(
-      list,
-      (e) => (e as Comparable).compareTo(value) <= 0,
-      start: start,
-      count: count,
-    );
-  } else {
-    return lowerBoundBy<E>(
-      list,
-      (e) => compare(e, value) <= 0,
-      start: start,
-      count: count,
-    );
+  int i, j, l, r, m, c;
+  final int n = list.length;
+
+  // determine range [i, j)
+  i = start ?? 0;
+  j = n;
+  if (count != null) {
+    if (count < 0) {
+      throw RangeError("count can not be negative");
+    }
+    j = i + count;
+    if (j > n) j = n;
   }
+  if (i < 0) i = 0;
+
+  l = i;
+  r = j;
+  if (compare == null) {
+    while (l < r) {
+      // the middle of the range
+      m = l + ((r - l) >> 1);
+
+      // compare middle item with value
+      c = (list[m] as Comparable).compareTo(value);
+
+      if (c <= 0) {
+        // middle item is lesser, select right range
+        l = m + 1;
+      } else {
+        // middle item is either equal or greater, select left range
+        r = m;
+      }
+    }
+  } else {
+    while (l < r) {
+      // the middle of the range
+      m = l + ((r - l) >> 1);
+
+      // compare middle item with value
+      c = compare(list[m], value);
+
+      if (c <= 0) {
+        // middle item is lesser, select right range
+        l = m + 1;
+      } else {
+        // middle item is either equal or greater, select left range
+        r = m;
+      }
+    }
+  }
+
+  if (l > j) l = j;
+
+  if (!exactMatch) {
+    // lower bound index
+    return l;
+  }
+
+  // prepare binary search result
+  l--;
+  if (i <= l) {
+    if (compare == null) {
+      if (list[l] == value) return l;
+    } else {
+      if (compare(list[l], value) == 0) return l;
+    }
+  }
+  return -1;
 }
 
 /// Returns the index of the _last_ occurance of the [value] in a sorted [list],
@@ -279,21 +268,18 @@ int upperBound<E, V>(
 ///
 /// * The [list] must be a sorted list of items, otherwise the behavior of this
 ///   method is not defined.
-/// * The [value] must be comparable with the list items. Otherwise,  -1 is returned.
+/// * The [value] must be comparable with the list items. Otherwise, [TypeError] is thrown.
 /// * If [start] is given, search start there and go towards the end of the [list].
 /// * If [start] is not below the length of the [list], -1 is returned.
 /// * If [start] is negative, search starts at [start] + [count] or 0, whichever is greater.
 /// * If the [count] parameter is given, it will check up to [count] numbers of items.
-/// * If [count] is negative, -1 is returned.
+/// * If [count] is negative, [RangeError] is thrown.
 /// * [compare] is a custom comparator function between a list element and the value.
 ///   If it is null, `compareTo` method of [list] item is used.
 ///
 /// ## Details
 ///
-/// Internally the [lowerBoundBy] is called comparing whether the list item is less
-/// than or equal to the value as the `test` function, and with a `testEqual` function
-/// to perform binrary search directly.
-/// On [TypeError] or [RangeError], -1 is returned.
+/// Internally it calls the [upperBound] method, passing true to the `exactMatch` parameter.
 ///
 /// -------------------------------------------------------------------------
 /// Complexity: Time `O(log n)` | Space `O(1)`
@@ -304,35 +290,12 @@ int binarySearchMax<E, V>(
   final int? count,
   final EntryComparator<E, V>? compare,
 }) {
-  try {
-    int x;
-    if (compare == null) {
-      x = lowerBoundBy<E>(
-        list,
-        (e) => (e as Comparable).compareTo(value) <= 0,
-        start: start,
-        count: count,
-      );
-    } else {
-      x = lowerBoundBy<E>(
-        list,
-        (e) => compare(e, value) <= 0,
-        start: start,
-        count: count,
-      );
-    }
-    if (x > 0 && x > (start ?? 0)) {
-      x--;
-      if (compare == null) {
-        if (list[x] == value) return x;
-      } else {
-        if (compare(list[x], value) == 0) return x;
-      }
-    }
-  } on RangeError {
-    // ignore: empty_catches
-  } on TypeError {
-    // ignore: empty_catches
-  }
-  return -1;
+  return upperBound<E, V>(
+    list,
+    value,
+    start: start,
+    count: count,
+    compare: compare,
+    exactMatch: true,
+  );
 }