+++ lib/src/utils/comparators.dart
@@ -9,23 +9,16 @@
 /// * positive integer if [left] is greater than [right]
 typedef EntryComparator<E, V> = int Function(E left, V right);
 
-/// A template for testing an item for equality.
+/// A template for test function for an element.
 ///
 /// The return value of this function should be:
 /// * `true` if the [element] is equal,
 /// * `false` otherwise.
 typedef EqualityTest<E> = bool Function(E element);
 
-/// A template for testing if an item is lesser.
+/// A template for test function for an element.
 ///
 /// The return value of this function should be:
 /// * `true` if the [element] is strictly lesser,
 /// * otherwise, `false` if equal or greater.
 typedef LessThanTest<E> = bool Function(E element);
-
-/// A template for testing if an item is greater.
-///
-/// The return value of this function should be:
-/// * `true` if the [element] is strictly greater,
-/// * otherwise, `false` if equal or greater.
-typedef GreaterThanTest<E> = bool Function(E element);